created a server.js file and split the api code for user readabilty and overall site file structure
